Delete Filesystem abstract class
