+++ Source/FileManager/Filesystem/File.php
@@ -2,6 +2,7 @@
 
 namespace Saeghe\Saeghe\FileManager\Filesystem;
 
+use Saeghe\Saeghe\FileManager\Path;
 use function Saeghe\Saeghe\FileManager\File\chmod;
 use function Saeghe\Saeghe\FileManager\File\content;
 use function Saeghe\Saeghe\FileManager\File\create;
@@ -12,8 +13,22 @@ use function Saeghe\Saeghe\FileManager\File\modify;
 use function Saeghe\Saeghe\FileManager\File\permission;
 use function Saeghe\Saeghe\FileManager\File\preserve_copy;
 
-class File extends Filesystem
+class File implements \Stringable
 {
+    use Address;
+
+    public function __construct(public Path $path) {}
+
+    public static function from_string(string $path_string): static
+    {
+        return new static(Path::from_string($path_string));
+    }
+
+    public function __toString(): string
+    {
+        return $this->path->string();
+    }
+
     public function chmod(int $permission): self
     {
         chmod($this->path, $permission);
