--- Source/FileManager/Filesystem/Filesystem.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace Saeghe\Saeghe\FileManager\Filesystem;
-
-use Saeghe\Saeghe\FileManager\Path;
-
-abstract class Filesystem implements \Stringable
-{
-    use Address;
-
-    public function __construct(public Path $path) {}
-
-    public static function from_string(string $path_string): static
-    {
-        return new static(Path::from_string($path_string));
-    }
-
-    public function __toString(): string
-    {
-        return $this->path->string();
-    }
-}
